Bug fixed: Nullpointer-Exception when trying to book on vacation suborder that was valid at day of timereport but not at day of booking
Bug fixed: when prolonging an employeecontract, the employees suborders were not prolonged as well. This resulted in the creation of new standard employeeorders with start=contractbegin when the employee logged in after old employeecontract-validity. This in turn resulted in an error when a timereport was attempted to book in the time where the original and the new employeeorder both were valid. Now all employeeorders are being prolonged when the employeecontract is prolonged (with the exception of employeeorders whose suborder-duration is shorter than contractduration)