#436 Update order type retrieval method to support overrides
Modified the `TimereportDAO` and `TimereportService` to use a new method `getEffectiveOrderType` which can override the order type at the suborder level. Added `getEffectiveOrderType` method in `Suborder` class to return the suborder's order type if available, otherwise falling back to the customer order's order type. This ensures flexibility and correctness in order type management.